+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 
+## 0.3.0 (December 2, 2021)
+
+### Release 0.3.0
+* adding prelude module.
+* please use<br>
+`pub use argone::{*, prelude::*};`<br>or<br>`pub use argone::{..., prelude::*};`
+
+---
+
 ## 0.2.0 (December 2, 2021)
 
 ### Release 0.2.0